backend/requirements.txt
# Voice Agent Platform - Backend Dependencies
# Production Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
httpx==0.25.2

# Firebase Integration
firebase-admin==6.4.0
google-cloud-firestore==2.13.1

# Development Dependencies  
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# External API Clients
elevenlabs==0.2.26
twilio==8.10.1

# Testing & Mocking
factory-boy==3.3.1
faker==20.1.0
responses==0.24.1